backend/requirements.txt
cryptography==41.0.2
Flask==2.3.2
openai==0.27.0
python-dotenv==1.0.0
requests==2.25.1